AuctionCreatedConsumer/requirements.txt
requests==2.26.0
apache-beam[gcp]==2.34.0
web3==5.25.0
psycopg2-binary==2.9.2
